# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Brownie API
  version: "0.1.0"
  description: >
    The public API surface under /api/v1. It includes the platform
    self-check routes and the implemented workspace document, job, and
    persisted progress-event routes.

    Three answers are described here once instead of under each route,
    all in the Error shape. Any route can answer 429 RATE_LIMITED with a
    Retry-After header, when one person (or, for someone not signed in,
    one address) has made more requests of one kind in a minute than
    allowed, and 503 DATABASE_UNAVAILABLE with a Retry-After header. Every
    route that looks up the person asking, which is all of them but
    /api/v1/capabilities and /api/v1/data-practices, can answer 401
    SESSION_NO_LONGER_VALID, when the session is genuine but the person it
    names has no identity record any more, in which case the session has
    been ended. A request body larger than the API reads, other than a
    file upload, is answered 413 CONTENT_TOO_LARGE whatever kind of body
    it says it is.
servers:
  - url: http://localhost:8081
    description: Local development
paths:
...
  /api/v1/workspaces/{workspaceId}/documents/{documentId}/assist/interpret:
    post:
      operationId: interpretAssistRequest
      summary: >
        Reads a typed request into one bounded Assist command and reports
        its scope (which field and what it holds now, or which finding)
        without doing anything. Free text that matches no command comes back
        as NONE with the list of what Brownie can do.
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistTextRequest"
      responses:
        "200":
          description: The interpretation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistInterpretationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/{workspaceId}/documents/{documentId}/assist/execute:
    post:
      operationId: executeAssistRequest
      summary: >
        Executes the interpreted command against the revision the person
        was looking at. A change or a rewrite returns a patch proposal for
        the ordinary accept route (nothing is applied here); an explanation
        returns text; a draft request returns DRAFT for the client to start
        extraction; a request that cannot be carried out is a 400 with the
        reason. Rewrite and explain make exactly one bounded model call.
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistExecuteRequest"
      responses:
        "200":
          description: The outcome.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistExecutionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "429":
          $ref: "#/components/responses/UsageLimitReached"
        "502":
          description: The one model call did not produce a usable answer; nothing on the document changed.
...
components:
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource is not visible in this workspace.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
    PreconditionFailed:
      description: The named approval or manifest is stale against the document's current revision.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
    UsageLimitReached:
      description: >
        A model allowance is used up (code USAGE_LIMIT_REACHED): what is
        left would not hold this request, or the first request of this
        run. Nothing was sent to the model and nothing was queued. The
        error's limit property says which allowance. A replay of a start
        that was accepted earlier is answered with that start instead.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    DocxAnchor:
      type: object
      description: >
        A place in one paragraph of a Word form, as the page layout gives it:
        the paragraph's nodeId and anchorTextHash, and start and end counted
        in Unicode code points into its anchor text (the text of its own
        runs, as the layout's anchorStart offsets count it). AT is a point
        (start equals end); REPLACE takes the place of the text between start
        and end; WHOLE_LINE replaces the whole line when it holds only a
        blank and otherwise goes at its end; EXISTING_CONTROL makes the
        control at controlNodeId the spot.
      required: [part, placement, parserVersion]
      properties:
        part:
          type: string
          enum: [MAIN_DOCUMENT, HEADER, FOOTER]
        paragraphNodeId:
          type: string
          nullable: true
        placement:
          type: string
          enum: [AT, REPLACE, WHOLE_LINE, EXISTING_CONTROL]
        start:
          type: integer
          minimum: 0
        end:
          type: integer
          minimum: 0
        anchorTextHash:
          type: string
          nullable: true
        parserVersion:
          type: string
        controlNodeId:
          type: string
          nullable: true
...
    Error:
      description: >
        The shape of every error response in this API, regardless of
        cause.
      type: object
      required: [status, title, code, correlationId, fields, recoveryActions]
      properties:
        status:
          type: integer
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Stable, machine-readable error code (e.g. NOT_FOUND, VALIDATION_FAILED).
        correlationId:
          type: string
          description: Echoes the X-Correlation-Id request/response header.
        fields:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        recoveryActions:
          type: array
          items:
            type: string
        limit:
          type: string
          enum: [WORKSPACE_MONTH, GLOBAL_MONTH]
          description: >
            Present only when code is USAGE_LIMIT_REACHED. WORKSPACE_MONTH
            is this workspace's own monthly allowance, which starts again
            next calendar month; GLOBAL_MONTH is the allowance every
            workspace shares, which only whoever runs the service can raise.
        reason:
          type: string
          description: >
            Present only for some codes: with CONNECTOR_RESOURCE_UNAVAILABLE,
            GONE or CANCELLED for an event, and GONE, ACCESS_LOST, TRASHED,
            DOWNLOAD_RESTRICTED or CHANGED_DURING_COPY for a Drive file; with
            CONNECTOR_RESOURCE_UNSUPPORTED, TYPE or NOT_UTF8 for a Drive file
            (a calendar entry has none); with CONNECTION_RECONNECT_REQUIRED,
            TOKEN_REJECTED, TOKEN_UNREADABLE or PERMISSION_MISSING; with
            CONNECTOR_RESOURCE_REFUSED, the copy's rejection code from the
            same checks an upload meets; with ACTION_NOT_PROPOSABLE, NO_EXPORT,
            EXPORT_STALE, FORMAT_NOT_EXPORTED, FILE_TOO_LARGE,
            HIDDEN_CHARACTERS, TIME_SKIPPED or INVALID; with
            UNSUPPORTED_MEDIA_TYPE, why the file was refused: NOT_A_DOCUMENT,
            SPREADSHEET, PRESENTATION, PASSWORD_PROTECTED, RIGHTS_PROTECTED,
            REMOTE_CONTENT (it asks whoever opens it to fetch something from a
            network) or DAMAGED; with FILLABLE_FORM_FAILED, CANNOT_OPEN, DAMAGED
            or TIMED_OUT.
        access:
          type: string
          enum: [DRIVE_FILES, CALENDAR_EVENTS, DRIVE_SAVING, CALENDAR_EVENT_CREATION]
          description: >
            Present only when code is CONNECTION_RECONNECT_REQUIRED, naming the connection that must be made
            again, or CONNECTION_ACCESS_NOT_OFFERED, naming the connection nothing here would use.
        actionType:
          type: string
          enum: [DRIVE_SAVE_FILE, DRIVE_SAVE_AS_GOOGLE_DOC, CALENDAR_CREATE_EVENT, GOOGLE_DOC_APPEND]
          description: Present only when code is ACTION_NOT_OFFERED, naming the kind of change not made here.
        problems:
          type: array
          description: >
            Present only when code is RULE_VALIDATION_FAILED: one entry per
            way the proposed rule's payload does not hold against the
            draft's own field definitions.
          items:
            type: object
            properties:
              reason:
                type: string
              detail:
                type: string
...
    PdfPageAnchor:
      type: object
      required: [kind, pageNumber]
      description: >
        A place selected on a PDF form's page: page pageNumber, and either a
        point (points, on the page as stored, before it is turned, origin at
        the top-left of its visible area, Y down: the convention boxes use)
        or one of the page's lines by its index in the page layout. Exactly
        one of point and lineIndex.
      properties:
        kind:
          type: string
          enum: [PDF]
        pageNumber:
          type: integer
          minimum: 1
        point:
          type: object
          required: [x, y]
          nullable: true
          properties:
            x:
              type: number
            y:
              type: number
        lineIndex:
          type: integer
          minimum: 0
          nullable: true
    AssistTextRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          maxLength: 1000
        pageAnchor:
          description: >
            The place selected on the page, if any: where "here" and "this
            line" are, or the choice picked from an interpretation's choices.
            A place in a Word form's text (DocxAnchor, with no kind) or on a
            PDF form's page (PdfPageAnchor, kind PDF).
          oneOf:
            - $ref: "#/components/schemas/DocxAnchor"
            - $ref: "#/components/schemas/PdfPageAnchor"
            - type: "null"
    AssistExecuteRequest:
      type: object
      required: [text, expectedRevisionId]
      properties:
        text:
          type: string
          maxLength: 1000
        expectedRevisionId:
          type: integer
          format: int64
        pageAnchor:
          oneOf:
            - $ref: "#/components/schemas/DocxAnchor"
            - $ref: "#/components/schemas/PdfPageAnchor"
            - type: "null"
    AssistPlaceChoiceResponse:
      type: object
      required: [lineText, anchor]
      properties:
        lineText:
          type: string
        anchor:
          description: >
            The place, in the shape the page sends it back in as pageAnchor:
            a place in a Word form's text, or a line on a PDF form's page
            (kind PDF, pageNumber and lineIndex).
          oneOf:
            - $ref: "#/components/schemas/DocxAnchor"
            - $ref: "#/components/schemas/PdfPageAnchor"
    AssistSpotChangeResponse:
      type: object
      required: [fieldId, label, previousRevisionId, templateVersionId]
      properties:
        fieldId:
          type: string
        label:
          type: string
        lineText:
          type: string
          nullable: true
          description: The line an added spot went on; null for a rename, a removal, or a box put at a point on a PDF page.
        previousRevisionId:
          type: integer
          format: int64
          description: The revision the change replaced; restoring it is Undo.
        templateVersionId:
          type: integer
          format: int64
    AssistScopeResponse:
      type: object
      properties:
        fieldId:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        currentValue:
          type: string
          nullable: true
        findingMessage:
          type: string
          nullable: true
    AssistInterpretationResponse:
      type: object
      required: [kind, summary, executable, usesModel, help]
      properties:
        kind:
          type: string
          enum: [DRAFT, CHANGE_FIELD, REWRITE_FIELD, EXPLAIN_FINDING, ADD_FILL_SPOT, RENAME_FILL_SPOT, REMOVE_FILL_SPOT, NONE]
        summary:
          type: string
        scope:
          allOf:
            - $ref: "#/components/schemas/AssistScopeResponse"
          nullable: true
        executable:
          type: boolean
        usesModel:
          type: boolean
        help:
          type: array
          items:
            type: string
        choices:
          type: array
          description: >
            The lines a new fill spot could go on when the quoted words are on
            more than one; send the chosen anchor back as pageAnchor with the
            same text. Empty otherwise.
          items:
            $ref: "#/components/schemas/AssistPlaceChoiceResponse"
    AssistExecutionResponse:
      type: object
      required: [kind, summary, help]
      properties:
        kind:
          type: string
          enum: [DRAFT, CHANGE_FIELD, REWRITE_FIELD, EXPLAIN_FINDING, ADD_FILL_SPOT, RENAME_FILL_SPOT, REMOVE_FILL_SPOT, NONE]
        summary:
          type: string
        proposal:
          allOf:
            - $ref: "#/components/schemas/PatchProposalResponse"
          nullable: true
        explanation:
          type: string
          nullable: true
        help:
          type: array
          items:
            type: string
        spotChange:
          allOf:
            - $ref: "#/components/schemas/AssistSpotChangeResponse"
          nullable: true
          description: Set when a fill spot was added, renamed or taken away, which is applied at once.
...
    ProposedFieldResponse:
      type: object
      description: >
        Exactly one of value (SCALAR) or values (REPEATED) is set, the
        same shape FieldValueResponse uses for a stored field.
      required: [type, cardinality, evidenceSpanIds]
      properties:
        type:
          type: string
          enum: [TEXT, DATE]
        cardinality:
          type: string
          enum: [SCALAR, REPEATED]
        value:
          type: [string, "null"]
        values:
          type: [array, "null"]
          items:
            type: string
        evidenceSpanIds:
          type: array
          items:
            type: integer
            format: int64
    SkippedRepeatedItemResponse:
      type: object
      description: >
        A repeated row (an action item, in the built-in templates) the
        extraction result contained but the proposal could not carry:
        a row is only proposed whole, and the content model has no slot
        for an unknown owner or due date. Reported by name rather than
        dropped silently.
      required: [itemIndex, unresolvedFieldIds]
      properties:
        itemIndex:
          type: integer
          description: The row's zero-based position in the extraction result.
        unresolvedFieldIds:
          type: array
          items:
            type: string
        description:
          type: [string, "null"]
          description: The row's first usable text, typically the task itself, bounded in length.
    PatchProposalResponse:
      type: object
      required: [id, documentId, baseRevisionId, proposedValues, status, createdAt, proposedRepeatedItemCount, skippedRepeatedItems]
      properties:
        id:
          type: integer
          format: int64
        documentId:
          type: integer
          format: int64
        baseRevisionId:
          type: integer
          format: int64
        proposedValues:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ProposedFieldResponse"
        status:
          type: string
          enum: [PROPOSED, ACCEPTED]
        createdAt:
          type: string
          format: date-time
        proposedRepeatedItemCount:
          type: integer
          description: How many complete repeated rows the proposal carries.
        skippedRepeatedItems:
          type: array
          items:
            $ref: "#/components/schemas/SkippedRepeatedItemResponse"
...